upm/requirements.txt
# Core Application Dependencies
# These packages are required for the application to function correctly.

# Core Utilities
pyyaml==6.0.2
requests==2.32.4
packaging==25.0
tenacity==9.1.2
pydantic==2.8.2

# CLI & TUI Features
typer==0.16.0
fastapi==0.116.1
uvicorn[standard]==0.35.0
textual==4.0.0
colorama==0.4.6

# Security & Encryption (Required)
cryptography==45.0.5

# AI & Machine Learning Features
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.26.2
pandas==2.1.3
shap==0.45.1

# Ecosystem Specific
toml==0.10.2
slack_sdk==3.36.0
boto3==1.35.0
distro==1.9.0